Start testing

+++ src/commonInterface/AbstractRobot.java
@@ -30,11 +30,11 @@ public abstract class AbstractRobot {
 		return false;
 	}
 	
-	public int getV1() {
+	public double getV1() {
 		return robotMotors.get(0).getSpeed();
 	}
 	
-	public int getV2() {
+	public double getV2() {
 		return robotMotors.get(1).getSpeed();
 	}
 }
